scrollTopButton component moved to common components
