+++ src/pages/home/Home.js
@@ -10,7 +10,7 @@ import LegalModal from './components/LegalModal';
 import Hero from './components/Hero';
 import HomeSection from './components/HomeSection';
 import SocialMedia from './components/SocialMedia';
-import ScrollTopButton from './components/ScrollTopButton';
+import ScrollTopButton from 'components/scrollTopButton/ScrollTopButton';
 import ImageAbout from './assets/images/about.svg';
 import ImageMission from './assets/images/mission.svg';
 import ImageContact from './assets/images/contact.svg';
